Performance optimizations
* Do typecast on array instead of individual elements
* Change constant types to match the type we need during serialization
* Tighten return types of functions to reduce type conversions
* Preallocate common serialization prefix
* Inline getTimeID for cases where we'll most likely hit the cache
* Change branch order to check the most frequent case first

These reduce the serialization time by ~20% and reduce the total sequence creationg+generation time by ~6%.

+++ ExpSeq.m
@@ -242,16 +242,16 @@ classdef ExpSeq < RootSeq
                     assert(defval);
                     defval_serialized = [defval_serialized, ...
                                          typecast(cid, 'int8'), ...
-                                         int8(SeqVal.TypeBool), int8(1)];
+                                         SeqVal.TypeBool, int8(1)];
                 elseif isa(defval, 'int32')
                     defval_serialized = [defval_serialized, ...
                                          typecast(cid, 'int8'), ...
-                                         int8(SeqVal.TypeInt32), typecast(defval, 'int8')];
+                                         SeqVal.TypeInt32, typecast(defval, 'int8')];
                 else
                     defval = double(defval);
                     defval_serialized = [defval_serialized, ...
                                          typecast(cid, 'int8'), ...
-                                         int8(SeqVal.TypeFloat64), typecast(defval, 'int8')];
+                                         SeqVal.TypeFloat64, typecast(defval, 'int8')];
                 end
                 ndefvals = ndefvals + 1;
             end
